change somthing in  design in dashboard and andel links in profile and move the users api to app and test it, all works without errors.
